impl bukkit

+++ games-core/src/main/java/us/smartmc/gamescore/instance/serialization/CuboidWrapper.java
@@ -1,7 +1,7 @@
 package us.smartmc.gamescore.instance.serialization;
 
 import lombok.Getter;
-import org.joml.Vector3d;
+import org.joml.Vector3i;
 
 import java.io.Serializable;
 import java.util.List;
@@ -10,10 +10,10 @@ import java.util.List;
 public class CuboidWrapper implements Serializable {
 
     private final List<BlockStateWrapper> blocks;
-    private final Vector3d min, max;
+    private final Vector3i min, max;
 
     // Constructor que toma un cuboide y serializa los bloques dentro de la región
-    public CuboidWrapper(Vector3d min, Vector3d max, List<BlockStateWrapper> blocks) {
+    public CuboidWrapper(Vector3i min, Vector3i max, List<BlockStateWrapper> blocks) {
         this.min = min;
         this.max = max;
         this.blocks = blocks;
